chore(build): don't rename file if it does not exist
See commit fba49b9408d2e6260b7eaf9ca93a52e018ba4ad2

+++ scripts/build_locally.sh
@@ -692,7 +692,9 @@ pushd "${BUILD_DIR}" > /dev/null
     ${CONAN_STATIC_BUILD_FLAGS} \
     --build missing \
 
-  mv "${BUILD_DIR}/Findghc_filesystem.cmake" "${BUILD_DIR}/FindghcFilesystem.cmake"
+  if [ -f "${BUILD_DIR}/Findghc_filesystem.cmake" ]; then
+    mv "${BUILD_DIR}/Findghc_filesystem.cmake" "${BUILD_DIR}/FindghcFilesystem.cmake"
+  fi
 
   print 57 "Generate source files";
   python3 "${PROJECT_ROOT_DIR}/packages/nextclade_common/scripts/generate_cli.py" \
